Improved OpenNLP memory usage by intern()'ing model strings.
OpenNLP 1.5 api doesn't support PooledGISModelReader using the new API. Provided variants of GenericModelSerializer/Reader and TokenNameFinderModel that now intern() strings, which is useful when loading multiple name finder models that have significant overlap.

Added tests to NameFinderTest that measure the memory used for each approach with the aid of the MemoryStatus class, which reads and totals memory usage according to the memory-related MBeans exposed by the java virtual machine.

Tests run locally show that when the english person, location and date maxent/gis models are loaded with the non-pooled model reader, approximately 150M is consumed, while only approximately 100M is used when the same models are loaded using the pooled model readers.